modal is populated with location specific restaurants

+++ models/user.js
@@ -1,7 +1,7 @@
 module.exports = function(sequelize, DataTypes) {
   var User = sequelize.define("User", {
     name: DataTypes.STRING,
-    cell: DataTypes.INTEGER
+    cell: DataTypes.BIGINT
   });
 
   User.associate = function(models) {
